Removed old obsolete functional tests (moved to internal modules)

+++ core/colors.hpp
@@ -53,7 +53,7 @@ typedef RGBcolor RGBPalette[256];
 static inline RGBcolor color_decode(const RGBcolor c, const uint8_t in_bpp, const uint32_t (& palette)[256]){
     switch (in_bpp){
     case 8:
-      return palette[c] & 0xFFFFFF;
+      return palette[(uint8_t)c] & 0xFFFFFF;
     case 15:
     {
         // r1 r2 r3 r4 r5 g1 g2 g3 g4 g5 b1 b2 b3 b4 b5
